Refactor BatchPreparedStatementExecutorTest
